refactor post processor, delete ensure_imported
